pyproject.toml
[build-system]
requires = ["flit-core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "friendlydateparser"
version = "0.2.0"
description = "A flexible date parsing library that can handle complex natural language expressions, making date manipulation easy and intuitive."
readme = "README.md"
authors = [ { name = "Salvador Fandiño García", email = "sfandino@yahoo.com" } ]
license = { text = "MIT" }
dependencies = [ "antlr4-python3-runtime >= 4.13",
                 "python-dateutil >= 2.8",
                 "pytz >= 2024", ]
keywords = [ "date parsing", "natural language", "calendar", "flexible dates", "utility", "date", "parser" ]

[project.urls]
Source = "https://github.com/salva/py-friendlydateparser"

[tool.flit.sdist]
include = [
    "src/friendlydateparser/antlr/FriendlyDateLexer.py",
    "src/friendlydateparser/antlr/FriendlyDateParser.py",
    "src/friendlydateparser/antlr/FriendlyDateVisitor.py",
]
